+++ Assets/FactoryCoreLogic/Items/AnthraciteCoal.cs
@@ -5,7 +5,7 @@ namespace Core
         public override ItemType Type => ItemType.AnthraciteCoal;
         public override uint MaxStack => 50_000_000;
         public override UnitType Units => UnitType.Milligram;
-        public AnthraciteCoal(uint quantity) : base(quantity) { }
+        public AnthraciteCoal(ulong quantity) : base(quantity) { }
         override public float Width => .3f;
         private const string name = "Anthracite coal";
         public override string Name => name;
