+++ Assets/FactoryCoreLogic/Items/Chalcopyrite.cs
@@ -9,6 +9,6 @@ namespace Core
         public override float? MeltingPoint_Celsious => 950;
         public override uint MaxStack => 200_000_000;
         public override UnitType Units => UnitType.Milligram;
-        public Chalcopyrite(uint quantity) : base(quantity) { }
+        public Chalcopyrite(ulong quantity) : base(quantity) { }
     }
 }
