+++ Assets/FactoryCoreLogic/Items/Dirt.cs
@@ -4,7 +4,7 @@ namespace Core
     {
         public override ItemType Type => ItemType.Dirt;
         public override uint MaxStack => 4;
-        public Dirt(uint quantity) : base(quantity) { }
+        public Dirt(ulong quantity) : base(quantity) { }
         private const string name = "Dirt";
         public override string Name => name;
         public override string? ChemicalFormula => null;
