+++ Assets/FactoryCoreLogic/Items/IronSiliconSlag.cs
@@ -7,6 +7,6 @@ namespace Core
         public override string? ChemicalFormula => "FeS₂";
         public override uint MaxStack => 200_000_000;
         public override UnitType Units => UnitType.Milligram;
-        public IronSiliconSlag(uint quantity) : base(quantity) { }
+        public IronSiliconSlag(ulong quantity) : base(quantity) { }
     }
 }
