+++ Assets/FactoryCoreLogic/Items/Item.schema.cs
@@ -1,4 +1,7 @@
+using System;
+using Core;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace Schema
 {
@@ -8,7 +11,7 @@ namespace Schema
         public Core.ItemType Type { get; set; }
 
         [JsonProperty("quantity")]
-        public uint Quantity { get; set; }
+        public ulong Quantity { get; set; }
 
         [JsonProperty("id")]
         public ulong Id { get; set; }
