+++ Assets/FactoryCoreLogic/Items/Log.cs
@@ -11,6 +11,6 @@ namespace Core
         public override string Name => name;
         public override string? ChemicalFormula => "C₆H₁₀O₅";
 
-        public Log(uint quantity) : base(quantity) { }
+        public Log(ulong quantity) : base(quantity) { }
     }
 }
