+++ Assets/FactoryCoreLogic/Items/Wood.cs
@@ -4,7 +4,7 @@ namespace Core
     {
         public override ItemType Type => ItemType.Wood;
         public override uint MaxStack => 4;
-        public Wood(uint quantity) : base(quantity) { }
+        public Wood(ulong quantity) : base(quantity) { }
         private const string name = "Wood";
         public override string Name => name;
         public override string? ChemicalFormula => "C₆H₁₀O₅";
